more frontend

+++ app/services/pinecone_service.py
@@ -54,5 +54,5 @@ def get_most_similar_chunks_for_query(query, index_name):
     return context_chunks   
 
 def delete_index(index_name):
+  if index_name in pinecone.list_indexes():
     pc.delete_index(name=index_name)
-    print(f"Index {index_name} deleted successfully")
